requirements.txt
annotated-types==0.7.0
cysystemd==2.0.1
evdev==1.9.2
pydantic==2.11.4
pydantic-core==2.33.2
pyyaml==6.0.2
systemd==0.17.1
typing-extensions==4.13.2
typing-inspection==0.4.0
